requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
line-bot-sdk
httpx
beautifulsoup4
python-dotenv
pydantic-settings
